Move egui state structs to dedicated module
