+++ crates/fj-viewer/src/graphics/renderer.rs
@@ -10,6 +10,7 @@ use wgpu_glyph::ab_glyph::InvalidFont;
 
 use crate::{
     camera::Camera,
+    gui::EguiState,
     screen::{Screen, Size},
 };
 
@@ -18,28 +19,6 @@ use super::{
     pipelines::Pipelines, transform::Transform, uniforms::Uniforms,
     vertices::Vertices, DEPTH_FORMAT,
 };
-
-#[derive(Default)]
-struct EguiOptionsState {
-    show_trace: bool,
-    show_layout_debug_on_hover: bool,
-    show_debug_text_example: bool,
-    show_settings_ui: bool,
-    show_inspection_ui: bool,
-}
-
-pub struct EguiState {
-    pub winit_state: egui_winit::State,
-    pub context: egui::Context,
-    rpass: egui_wgpu::renderer::RenderPass,
-    options: EguiOptionsState,
-}
-
-impl std::fmt::Debug for EguiState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("EguiState {}")
-    }
-}
 
 /// Graphics rendering state and target abstraction
 #[derive(Debug)]
