+++ crates/fj-viewer/src/lib.rs
@@ -18,3 +18,5 @@ pub mod camera;
 pub mod graphics;
 pub mod input;
 pub mod screen;
+
+mod gui;
